+++ examples/scripts/thermal_lithium_ion.py
@@ -18,7 +18,7 @@ models = [full_thermal_model, lumped_thermal_model]
 
 # load parameter values and process models and geometry
 param = models[0].default_parameter_values
-param.update({"Heat transfer coefficient [W.m-2.K-1]": 0.1})
+param.update({"Heat transfer coefficient [W.m-2.K-1]": 1})
 
 for model in models:
     param.process_model(model)
@@ -40,7 +40,8 @@ for model in models:
 solutions = [None] * len(models)
 t_eval = np.linspace(0, 0.17, 100)
 for i, model in enumerate(models):
-    solution = model.default_solver.solve(model, t_eval)
+    solver = pybamm.ScipySolver(atol=1e-8, rtol=1e-8)
+    solution = solver.solve(model, t_eval)
     solutions[i] = solution
 
 # plot